Moved installing of embedded packages to PluginAssemblyStore
