+++ framework/OpenMod.Core/Plugins/PluginAssemblyStore.cs
@@ -1,4 +1,11 @@
-﻿using System;
+﻿using Microsoft.Extensions.Logging;
+using NuGet.Packaging.Core;
+using NuGet.Versioning;
+using OpenMod.API;
+using OpenMod.API.Plugins;
+using OpenMod.NuGet;
+using Semver;
+using System;
 using System.Collections.Generic;
 using System.Drawing;
 using System.IO;
@@ -6,11 +13,8 @@ using System.Linq;
 using System.Reflection;
 using System.Text.RegularExpressions;
 using System.Threading.Tasks;
-using Microsoft.Extensions.Logging;
-using OpenMod.API;
-using OpenMod.API.Plugins;
-using OpenMod.NuGet;
-using Semver;
+using YamlDotNet.Serialization;
+using YamlDotNet.Serialization.NamingConventions;
 
 namespace OpenMod.Core.Plugins
 {
@@ -46,6 +50,44 @@ namespace OpenMod.Core.Plugins
             }
         }
 
+        private async Task InstallPackagesInEmbeddedFile(Assembly assembly)
+        {
+            var deserializer = new DeserializerBuilder()
+                .WithNamingConvention(CamelCaseNamingConvention.Instance)
+                .Build();
+            
+            try
+            {
+                var packagesResourceName = assembly.GetManifestResourceNames()
+                    .FirstOrDefault(x => x.EndsWith("packages.yaml"));
+                if (packagesResourceName == null) return;
+
+                using var stream = assembly.GetManifestResourceStream(packagesResourceName);
+                if (stream == null) return;
+
+                using var reader = new StreamReader(stream);
+                var packagesContent = await reader.ReadToEndAsync();
+
+                var deserialized = deserializer.Deserialize<SerializedPackagesFile>(packagesContent).Packages;
+
+                var packages = deserialized?.Select(d => new PackageIdentity(d.Id,
+                        d.Version.Equals("latest", StringComparison.OrdinalIgnoreCase)
+                            ? null
+                            : new NuGetVersion(d.Version)))
+                    .ToList();
+
+                if (packages == null || packages.Count == 0) return;
+
+                m_Logger.LogInformation($"Found and installing embedded NuGet packages for plugin assembly: {assembly.GetName().Name}");
+
+                await m_NuGetPackageManager.InstallPackagesAsync(packages);
+            }
+            catch (Exception ex)
+            {
+                m_Logger.LogError(ex, $"Failed to check/load embedded NuGet packages for assembly: {assembly}");
+            }
+        }
+
         public async Task<ICollection<Assembly>> LoadPluginAssembliesAsync(IPluginAssembliesSource source)
         {
             if (source == null)
@@ -64,6 +106,8 @@ namespace OpenMod.Core.Plugins
                     providerAssemblies.Remove(providerAssembly);
                     continue;
                 }
+
+                await InstallPackagesInEmbeddedFile(providerAssembly);
 
                 ICollection<Type> types;
                 try
