+++ framework/OpenMod.Runtime/OpenModStartup.cs
@@ -3,8 +3,6 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Logging;
 using NuGet.Packaging;
-using NuGet.Packaging.Core;
-using NuGet.Versioning;
 using OpenMod.API;
 using OpenMod.API.Ioc;
 using OpenMod.API.Plugins;
@@ -22,8 +20,6 @@ using System.IO;
 using System.Linq;
 using System.Reflection;
 using System.Threading.Tasks;
-using YamlDotNet.Serialization;
-using YamlDotNet.Serialization.NamingConventions;
 
 namespace OpenMod.Runtime
 {
@@ -100,46 +96,7 @@ namespace OpenMod.Runtime
                     m_Logger.LogError(ex, $"Failed to copy resources from assembly: {assembly}");
                 }
             }
-
-            var deserializer = new DeserializerBuilder()
-                .WithNamingConvention(CamelCaseNamingConvention.Instance)
-                .Build();
-
-            foreach (var assembly in assemblies)
-            {
-                // Install packages specified in packages.yaml
-                try
-                {
-                    var packagesResourceName = assembly.GetManifestResourceNames()
-                        .FirstOrDefault(x => x.EndsWith("packages.yaml"));
-                    if (packagesResourceName == null) continue;
-
-                    using var stream = assembly.GetManifestResourceStream(packagesResourceName);
-                    if (stream == null) continue;
-
-                    using var reader = new StreamReader(stream);
-                    var packagesContent = await reader.ReadToEndAsync();
-
-                    var deserialized = deserializer.Deserialize<SerializedPackagesFile>(packagesContent).Packages;
-                    
-                    var packages = deserialized?.Select(d => new PackageIdentity(d.Id,
-                        d.Version.Equals("latest", StringComparison.OrdinalIgnoreCase)
-                            ? null
-                            : new NuGetVersion(d.Version)))
-                        .ToList();
-
-                    if (packages == null || packages.Count == 0) continue;
-
-                    m_Logger.LogInformation($"Found and installing embedded NuGet packages for plugin assembly: {assembly.GetName().Name}");
-
-                    await m_NuGetPackageManager.InstallPackagesAsync(packages);
-                }
-                catch (Exception ex)
-                {
-                    m_Logger.LogError(ex, $"Failed to check/load embedded NuGet packages for assembly: {assembly}");
-                }
-            }
-
+            
             foreach (var assembly in assemblies)
             {
                 // Auto register services with [Service] and [ServiceImplementation] attributes
